+++ template/component.h.erb
@@ -64,6 +64,7 @@ struct el {
 struct common_data_s {
   name_to_dispatcher_t *name_to_dispatcher;
   static_callbacks_t *static_callbacks;
+  void *on_push_callback;
 
   void *usr_data;
   const bt_value *params;
@@ -104,6 +105,7 @@ typedef enum btx_source_status_e btx_source_status_t;
 struct common_data_s {
   name_to_dispatcher_t *name_to_dispatcher;
   static_callbacks_t *static_callbacks;
+  void *on_push_callback;
 
   void *usr_data;
   const bt_value *params;
