+++ template/downstream.c.erb
@@ -20,9 +20,9 @@ void btx_downstream_move_messages(
   }
 }
 
-void btx_downstream_push_message(
-    btx_message_iterator_t *message_iterator_private_data,
-    const bt_message *message) {
+static void
+_btx_push_message(btx_message_iterator_t *message_iterator_private_data,
+                  const bt_message *message) {
 
   struct el *elt;
   if (message_iterator_private_data->pool) {
@@ -35,6 +35,31 @@ void btx_downstream_push_message(
   DL_APPEND(message_iterator_private_data->queue, elt);
 }
 
+// Exposed to the user
+void btx_push_message(void *btx_handle, const bt_message *message) {
+  _btx_push_message((btx_message_iterator_t *)btx_handle, message);
+}
+
+// User by everything
+void btx_downstream_push_message(
+    btx_message_iterator_t *message_iterator_private_data,
+    const bt_message *message) {
+
+  on_push_callback_f *p =
+      message_iterator_private_data->common_data->on_push_callback;
+  if (p)
+    p(((void *)message_iterator_private_data),
+      message_iterator_private_data->common_data->usr_data, message);
+  else
+    _btx_push_message(message_iterator_private_data, message);
+}
+
+// Register callbacks
+
+void btx_register_on_push_callback(void *btx_handle,
+                                   on_push_callback_f *callback) {
+  ((common_data_t *)btx_handle)->on_push_callback = callback;
+}
 bt_trace_class *
 btx_downstream_trace_class_create_rec(bt_self_component *_self_component) {
   bt_trace_class *_trace_class;
