+++ template/downstream.h.erb
@@ -32,6 +32,16 @@ void btx_push_message_<%= e.name_sanitized %>(
     void *
         btx_handle<%= ', int64_t _timestamp' if e.default_clock_class %><%= e.args.map{ |s| "#{s.type} #{s.name}" }.join_with_prefix(", ") %>);
 <% end %>
+
+void btx_push_message(void *btx_handle, const bt_message *message);
+
+// Downstream callbacks
+
+typedef void on_push_callback_f(void *btx_handle, void *usr_data,
+                                const bt_message *message);
+void btx_register_on_push_callback(void *btx_handle,
+                                   on_push_callback_f *callback);
+
 #ifdef __cplusplus
 }
 #endif
