+++ test/callbacks/on_push_callback/1.btx_model.yaml
@@ -1,0 +1,6 @@
+---
+:stream_classes:
+- :name: sc
+  :default_clock_class: {}
+  :event_classes:
+  - :name: event
